init 3895

+++ bmo-serveur/src/main/java/akandan/bahou/kassy/serveur/dao/InterfaceUtilisateurDAO.java
@@ -1,15 +1,15 @@
 package akandan.bahou.kassy.serveur.dao;
 
+import akandan.bahou.kassy.commun.util.ExceptionPersistance;
+import akandan.bahou.kassy.serveur.modele.Utilisateur;
 import java.util.List;
 import java.util.Optional;
-import akandan.bahou.kassy.serveur.modele.Utilisateur;
-import akandan.bahou.kassy.commun.util.ExceptionPersistance;
 
 public interface InterfaceUtilisateurDAO {
 
     Utilisateur creer(Utilisateur utilisateurACreer) throws ExceptionPersistance;
 
-    Optional<Utilisateur> trouverParId(int idUtilisateur) throws ExceptionPersistance;
+    Optional<Utilisateur> trouverParId(long idUtilisateur) throws ExceptionPersistance; // Corrigé en long
 
     Optional<Utilisateur> trouverParIdentifiantConnexion(String identifiantConnexion) throws ExceptionPersistance;
 
@@ -17,7 +17,7 @@ public interface InterfaceUtilisateurDAO {
 
     boolean mettreAJour(Utilisateur utilisateurAMettreAJour) throws ExceptionPersistance;
 
-    boolean supprimer(int idUtilisateur) throws ExceptionPersistance;
+    boolean supprimer(long idUtilisateur) throws ExceptionPersistance; // Corrigé en long
 
     boolean existeParIdentifiantConnexion(String identifiantConnexion) throws ExceptionPersistance;
 }
